migrate modal to insomnia-data
