+++ packages/insomnia/src/insomnia-data/node-src/services/websocket-payload.ts
@@ -1,0 +1,43 @@
+import { replaceIdsInFields } from '~/models/helpers/replace-ids-in-fields';
+
+import { database } from '../../src/database';
+import { models } from '../../src/models';
+import { type WebSocketPayload } from '../../src/models/types';
+
+const { type, name } = models.webSocketPayload;
+
+export const create = (patch: Partial<WebSocketPayload> = {}) => {
+  if (!patch.parentId) {
+    throw new Error(`New WebSocketPayload missing \`parentId\`: ${JSON.stringify(patch)}`);
+  }
+
+  return database.docCreate<WebSocketPayload>(type, patch);
+};
+
+export const remove = (obj: WebSocketPayload) => database.remove(obj);
+
+export const update = (obj: WebSocketPayload, patch: Partial<WebSocketPayload> = {}) => database.docUpdate(obj, patch);
+
+export async function duplicate(request: WebSocketPayload, patch: Partial<WebSocketPayload> = {}) {
+  // Only set name and "(Copy)" if the patch does
+  // not define it and the request itself has a name.
+  // Otherwise leave it blank so the request URL can
+  // fill it in automatically.
+  if (!patch.name && request.name) {
+    patch.name = `${request.name} (Copy)`;
+  }
+
+  return database.duplicate<WebSocketPayload>(request, {
+    name,
+    ...patch,
+  });
+}
+
+export const getById = (_id: string) => database.findOne<WebSocketPayload>(type, { _id });
+export const getByParentId = (parentId: string) => database.findOne<WebSocketPayload>(type, { parentId });
+
+export const all = () => database.find<WebSocketPayload>(type);
+
+export function rewriteReferences(payload: WebSocketPayload, idMapping: Map<string, string>): WebSocketPayload {
+  return { ...payload, ...replaceIdsInFields(payload, ['value'], idMapping) };
+}
