+++ packages/insomnia/src/insomnia-data/src/models/types.ts
@@ -61,3 +61,9 @@ export type { MockServer } from './mock-server';
 export type { UnitTest } from './unit-test';
 export type { UnitTestResult } from './unit-test-result';
 export type { UnitTestSuite } from './unit-test-suite';
+export type { SocketIOPayload } from './socket-io-payload';
+export type { SocketIORequest } from './socket-io-request';
+export type { SocketIOResponse } from './socket-io-response';
+export type { WebSocketPayload } from './websocket-payload';
+export type { WebSocketRequest } from './websocket-request';
+export type { WebSocketResponse } from './websocket-response';
