+++ packages/insomnia/src/insomnia-data/src/models/websocket-request.ts
@@ -1,7 +1,6 @@
-import { database } from '../common/database';
-import { replaceIdsInFields } from './helpers/replace-ids-in-fields';
-import type { RequestAuthentication, RequestHeader, RequestParameter, RequestPathParameter } from './request';
-import type { BaseModel } from './types';
+import { replaceIdsInFields } from '~/models/helpers/replace-ids-in-fields';
+import type { RequestAuthentication, RequestHeader, RequestParameter, RequestPathParameter } from '~/models/request';
+import type { BaseModel } from '~/models/types';
 
 export const name = 'WebSocket Request';
 
@@ -55,55 +54,9 @@ export const init = (): BaseWebSocketRequest => ({
 
 export const migrate = (doc: WebSocketRequest) => doc;
 
-export const create = (patch: Partial<WebSocketRequest> = {}) => {
-  if (!patch.parentId) {
-    throw new Error(`New WebSocketRequest missing \`parentId\`: ${JSON.stringify(patch)}`);
-  }
-
-  return database.docCreate<WebSocketRequest>(type, patch);
-};
-
-export const remove = (obj: WebSocketRequest) => database.remove(obj);
-
-export const update = (obj: WebSocketRequest, patch: Partial<WebSocketRequest> = {}) => database.docUpdate(obj, patch);
-
-// This is duplicated (lol) from models/request.js
-export async function duplicate(request: WebSocketRequest, patch: Partial<WebSocketRequest> = {}) {
-  // Only set name and "(Copy)" if the patch does
-  // not define it and the request itself has a name.
-  // Otherwise leave it blank so the request URL can
-  // fill it in automatically.
-  if (!patch.name && request.name) {
-    patch.name = `${request.name} (Copy)`;
-  }
-
-  // Get sort key of next request
-  const q = {
-    metaSortKey: {
-      $gt: request.metaSortKey,
-    },
-  };
-
-  const [nextRequest] = await database.find<WebSocketRequest>(type, q, {
-    metaSortKey: 1,
-  });
-  const nextSortKey = nextRequest ? nextRequest.metaSortKey : request.metaSortKey + 100;
-  // Calculate new sort key
-  const sortKeyIncrement = (nextSortKey - request.metaSortKey) / 2;
-  const metaSortKey = request.metaSortKey + sortKeyIncrement;
-  return database.duplicate<WebSocketRequest>(request, {
-    name,
-    metaSortKey,
-    ...patch,
-  });
-}
-
-export const getById = (_id: string) => database.findOne<WebSocketRequest>(type, { _id });
-
-export const findByParentId = (parentId: string) => database.find<WebSocketRequest>(type, { parentId });
-
-export const all = () => database.find<WebSocketRequest>(type);
-
 export function rewriteReferences(request: WebSocketRequest, idMapping: Map<string, string>): WebSocketRequest {
-  return { ...request, ...replaceIdsInFields(request, ['url', 'headers', 'authentication', 'parameters', 'pathParameters'], idMapping) };
+  return {
+    ...request,
+    ...replaceIdsInFields(request, ['url', 'headers', 'authentication', 'parameters', 'pathParameters'], idMapping),
+  };
 }
